Move Spree::ProductDuplicator to Spree::Core namespace
